requirements.txt
flask
numpy
pillow
tensorflow-cpu
tf-keras
gunicorn
